Change the sample SERVER.crt generation to work
In order to be able to use this self-signed certificate, the CN must
match the hostname.  Since we're using localhost for now, match that in
the certificate generated.  Using CN for this is somewhat deprecated,
and we really should be using the appropriate extension.

However, for any practical use (across hosts), the server certificate
should be a legitimate certificate (probably from letsencrypt).

Signed-off-by: David Brown <david.brown@linaro.org>