Properly use Subject and Authority Ids
In order to use the SubjectKeyId and AuthorityKeyId, we need to extract
the public key from the CSR, and use that when signing the new
certificate.  Store the keyid in the database to be able to use in the
future.  With this change, it will be necessary to regenerate the CA.crt
certificate, as ones generated by previous versions of this code don't
have the Id.  This change also causes a database schema change.

Fixes #7.

Signed-off-by: David Brown <david.brown@linaro.org>