Split out client parts into run-client.sh
It should now be possible to run:

    ./run-server.sh

in one terminal window, and once the HTTPS server has started, run

    ./run-client.sh

in another window.  It is possible to run the client multiple times,
although the script will overwrite the USER.* files on each subsequent
run, losing the particulars of a given device key.

Signed-off-by: David Brown <david.brown@linaro.org>