Split invocation scripts to allow separate invocation.
First, create a `setup-ca.sh` script that is intended to be run once per
developer.  This will create the CA and SERVER key and certificates.
The script will exit with an error if either of these is present, and it
will have to be cleaned up manually.

Second, the `run-server.sh` script just invokes the linaroca server.  It
can be run multiple times, and will not create new keys each time.

Lastly, `run-client.sh` is changed to `new-device.sh`.  The keys are
placed in files named after the device ID, so that multiple devices can
be registered and experimented with.

Signed-off-by: David Brown <david.brown@linaro.org>