Move certificate files into 'certs' directory
Instead of cluttering the main directory with lots of certificate files,
move them into a `certs` directory.

This also moves much of the documentation from the README.md file into
the scripts that invoke the commands directly.  This should help keep
the documentation and the code from getting out of sync.

Signed-off-by: David Brown <david.brown@linaro.org>