Change CSR request to CBOR from json
Change the payload type of the CSR request from json to cbor.  This will
be easier to manage on embedded devices, and a bit more compact, as the
certificate will not have to be base64 encoded.

This changes the instructions in the `README.md`, the associated tools,
and the `new-device.sh` script.  The new device script will test with
this cbor-based protocol.

Signed-off-by: David Brown <david.brown@linaro.org>