Rename devclass to bootstrap
Change the name of the "devclass" key and cert generated to bootstrap.
This should make it a bit clearer what the purpose of this cert is for,
it is how the device initially bootstraps with the linaroca to get the
device-specific key.

The private key should still be kept private, but will need to be shared
with all devices (of a given type) that need to bootstrap.

Signed-off-by: David Brown <david.brown@linaro.org>