Allow hostname to be overridden
In several places, we use 'localhost' as the hostname, both when
listening for sockets and in various fields in certificates.
Unfortunately, doing so limits access to the CA server to also be from
localhost.

In order to the CA server to be tested from other devices, it is
necessary to set the hostname to a real DNS name that points to this
device.  This can be done by setting the environment variable HOSTNAME
to the FQDN of the host that will be running the server.  It is
important that this name actually be usable to access the server.  It is
fine if it is behind a NAT, as long as the device being tested can
access the server through that address.

Setting HOSTNAME before running the server will cause it to listen on
that address, rather than localhost, and accept incoming connections.

Signed-off-by: David Brown <david.brown@linaro.org>