Add MQTT server information to the cert response
Under the `[server]` section in the `.linaroca.toml` file, handle
entries for 'hubname' and 'mqttport'.  These will both be handed off in
the information packet to the client so that the client doesn't need
hardcoded information about the particular service being used.
Currently, we are still assuming Azure as the provider.

Signed-off-by: David Brown <david.brown@linaro.org>