Remove the hostname from the device cert
It is misleading to have the hostname that the CA server runs on on the
device cert. If it were actually checked, it would mean that the devices
could only be running on the server.

Change this to a sample vendor name, with the option to set to a vendor
name in the future.

Signed-off-by: David Brown <david.brown@linaro.org>