Rewrite how hsm module works
Instead of opening and closing the hsm module for every call, put the
api and session in a state variable, and use that to make the various
calls. This also makes cleanup much easier, and is overall much more of
a "Go" solution.

This currently also removes the global initialization attempt on the
hsm.  Since the use of the hsm is intended to be optional, this doesn't
make sense globally, and this will eventually be replaced by a provider
that returns a generalized interface implementing needed key operations
either on the hsm or on a local key file, depending on the arguments
given.

Signed-off-by: David Brown <david.brown@linaro.org>